clients/telegram_bot/requirements.txt
python-telegram-bot
aiohttp
python-dotenv
google-cloud-secret-manager
pytest
pytest-asyncio
